Launch and monitor Node.js.
You'll never remember what you went through to figure out that CMake was
not generating a universal binary, or rather, why Google Chrome refused
to load your plugin.

You can watch Google Chrome load its plugins.

    chrome --debug-plugin-loading > foo.txt 2>&1

It was using CFBundlePreflightExecutable to check to see if the plugin
bundle would load, and CFBundlePreflightExecutable was saying something
along the lines of.

    PluginLib::ReadWebPluginInfo bundle failed preflight: The bundle
    “Synapse Peer” couldn’t be loaded because it doesn’t contain a
    version for the current architecture.

`CMakeLists.txt` has been updated to produce a universal binary for
32bit and 64bin Intel Macs. The location of the directive matters. I'd
put it at first down with the `INSTALL` commands and it went unnoticed.

You'll recall when you wrote the `attendent` that you imagined that
someone unfamiliar with the reasoning behind it would find it very
fragile. Changing out one executable for another did illustrate how
brittle the code is, but I was able to recall, possibly by absorbing the
comments as I traced the code, all the many unknowns that the
`attendant` operates under.

It hung when the out-of-process server did not send a shutdown string
and port number through stdout, which was disheartening, but I do recall
saying that getting going was the tricky part. If the server is running,
and attendant is waiting on a port number or similar on standard out,
well, it's going to wait. If the server had crashed, then stdout would
be closed and recover would kick in.

I created `tracer.js`, a UDP server that listens for messages from the
plugin and writes them to standard out. Trying to debug with Xcode was
overwhelming. I plan on instrumenting the code, then hopefully I'll be
have the same sort of debug experience regardless of platform.

Not sure where `monitor.js` is going to live in the long term.

Closes #5.